progressutil: check for errors again when done
There's a race between the time we first check the error slice and when
we check for allDone, since a copyReader might finish and try to report
an error anywhere between this loop. Correct this by checking for errors
again before returning.

This is a stopgap - the real fix here is to refactor this a little more
invasively.