capnslog: implement logrus.StdLogger
This commit adds a Panicln method to the PackageLogger to make it
implement all of the golang/log logging methods and thus making it
implement with the logrus.StdLogger interface. I want this so that Dex
can accept loggers from packages that use capnslog, notably bridge. This
is the only missing method to fulfil the interface.