Add hidden go.mod for CI
We've already surpassed major version 1 in this repository, so we can't
add a go.mod without bumping the module path and requiring callers to
update:

    https://go.dev/ref/mod#non-module-compat

However, this repository is in deep maintenance mode and it doesn't
make sense to perturb callers in this way.  At the same time, we want
to be able to do test builds in CI, and GOPATH is no longer the way.

Add hidden go.mod and go.sum files, and have build.sh and test.sh symlink
them into place.  Put them in .gitignore so they're not accidentally
committed.